backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
httpx==0.27.2
PyJWT==2.9.0
